Fixed 1001 Tests Feedback
Made the tests run twice for each student to get the plain text feedback to provide to students along with the JSON encoded information. Alterations could be made to the test runner so that it is not required to run the tests twice but it's running python tests so it's a small enough impact to not worry.