Fixed Issue Parsing Blackboard
Prevents multiple headers from overriding. This is because some clever students in 2019s2a2-gradebook.zip have added their name in the same format as the submission information to their comments. As a sneaky fix, we're just saying that the first key/pair value in a submission zip is final. Ideally we would want to separate and escape student comments to prevent this from happening.